refactor(MyVideos): add MyVideoComponent
